pyproject.toml
[project]
name = "pytest-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "djangorestframework>=3.16.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
pythonpath = [
    "C:/Users/math/pytest-course"
]

